pyproject.toml
[project]
name = "efmtool_link"
version = "0.0.8"
authors = [
  { name="Axel von Kamp", email="vonkamp@mpi-magdeburg.mpg.de" },
]
description = "Exposes various efmtool functionality in Python."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = ["jpype1", "numpy", "scipy", "cobra>=0.26.3", "psutil"]

[build-system]
requires = ["setuptools", "install-jdk>=1.1", "jpype1"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/cnapy-org/efmtool_link"
Issues = "https://github.com/cnapy-org/efmtool_link/issues"
